※임시 : ctr-detail 에서 null 오류 발생 억제
(null 값인 파라메터가 없는 상태에서 null 오류 발생)

+++ analytics-server/src/main/java/org/fastcatsearch/analytics/analysis/calculator/NDaysClickKeywordHitCalculator.java
@@ -94,7 +94,9 @@ public class NDaysClickKeywordHitCalculator extends Calculator<ClickLog> {
 					
 					try {
 						logger.trace("click log copy {} -> {}", file.getAbsolutePath(), targetFilePath);
-						FileUtils.copyFile(file, new File(targetFilePath));
+						if(file.exists()) {
+							FileUtils.copyFile(file, new File(targetFilePath));
+						}
 					} catch (IOException e) {
 						logger.error("", e);
 					};
